:pencil2: presentation 패키지명 오타 수정
패키지명 오타 수정
