+++ src/main/kotlin/com/msg/gauth/domain/client/Client.kt
@@ -1,6 +1,6 @@
 package com.msg.gauth.domain.client
 
-import com.msg.gauth.domain.client.persentation.dto.request.ClientUpdateReqDto
+import com.msg.gauth.domain.client.presentation.dto.request.ClientUpdateReqDto
 import com.msg.gauth.domain.user.User
 import com.msg.gauth.global.entity.BaseIdEntity
 import javax.persistence.*
