+++ src/main/kotlin/com/msg/gauth/domain/client/presentation/dto/response/ClientDetailResDto.kt
@@ -1,15 +1,17 @@
-package com.msg.gauth.domain.client.persentation.dto.response
+package com.msg.gauth.domain.client.presentation.dto.response
 
 import com.msg.gauth.domain.client.Client
 
-data class ClientRegisterResDto(
+data class ClientDetailResDto(
+    val id: Long,
     val clientId: String,
     val clientSecret: String,
     val redirectUri: String,
     val serviceName: String,
     val serviceUri: String,
-) {
+){
     constructor(client: Client): this(
+        id = client.id,
         clientId = client.clientId,
         clientSecret = client.clientSecret,
         redirectUri = client.redirectUri,
