+++ src/main/kotlin/com/msg/gauth/domain/client/presentation/dto/response/ClientRegisterResDto.kt
@@ -1,16 +1,18 @@
-package com.msg.gauth.domain.client.persentation.dto.response
+package com.msg.gauth.domain.client.presentation.dto.response
 
 import com.msg.gauth.domain.client.Client
 
-data class SingleClientResDto(
-    val id: Long,
+data class ClientRegisterResDto(
     val clientId: String,
+    val clientSecret: String,
+    val redirectUri: String,
     val serviceName: String,
     val serviceUri: String,
 ) {
     constructor(client: Client): this(
-        id = client.id,
         clientId = client.clientId,
+        clientSecret = client.clientSecret,
+        redirectUri = client.redirectUri,
         serviceName = client.serviceName,
         serviceUri = client.serviceUri,
     )
