+++ src/main/kotlin/com/msg/gauth/domain/client/services/UpdateClientService.kt
@@ -1,7 +1,7 @@
 package com.msg.gauth.domain.client.services
 
 import com.msg.gauth.domain.client.exception.ClientNotFindException
-import com.msg.gauth.domain.client.persentation.dto.request.ClientUpdateReqDto
+import com.msg.gauth.domain.client.presentation.dto.request.ClientUpdateReqDto
 import com.msg.gauth.domain.client.repository.ClientRepository
 import com.msg.gauth.domain.user.utils.UserUtil
 import com.msg.gauth.global.annotation.service.TransactionalService
